requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25

# Cryptography
cryptography>=42.0.0

# Authentication
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4

# Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Rate Limiting
slowapi>=0.1.9

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
httpx>=0.26.0
